Fix: unit_attributes 테이블 power double형으로 변경

+++ wdam/src/main/java/com/back/wdam/entity/UnitAttributes.java
@@ -50,7 +50,7 @@ public class UnitAttributes {
 
     private String damageState;
 
-    private Integer power;
+    private Double power;
 
     private String powerDistribution;
